IT WORKS: traits over enums
instead of taking just a enum or trait based approach, it looks like I
can use both to cover what I need for dynamic dispatch.

Enums allow me to have an ergonomic interface for different datatypes,
and also limits the types to what the library implements (no surprises).

Traits allow for overloading functions. This is especially important for
`apply` functions, which have to know the primitive type of the
Datatype.

My strategy is to have a trait MyTrait<T>, and then implement specific
instances `MyTrait<i32>`, etc., for the enum. Then the implementation
instance can check that it's routing to the proper enum value with
primitive type, and return an error if the routing is done incorrectly.

e.g. the implementation for MyEnum of MyTrait<i32> will check that the
enum value being acted on at that moment is the right fit, a
`MyEnum::ValueI32(Buffer<i32>)`, and if it isn't (meaning that some
other enum value is trying to be matched on, it will return a runtime
error.

--
No easy way to maintain a heterogenous collection while also being able
to access the primitive type, unless it's enum. If using structs, there
will be a tendancy for the primitive type to become a bound on the
collection. So the enum works best for this. (Enums are good for things
only known at runtime).

And then the trait overlays this.